Move internal classes into the new `internal` package
