+++ plugin/src/main/kotlin/com/vanniktech/maven/publish/internal/central/EndOfBuildAction.kt
@@ -1,4 +1,4 @@
-package com.vanniktech.maven.publish.central
+package com.vanniktech.maven.publish.internal.central
 
 internal sealed interface EndOfBuildAction {
   val runAfterFailure: Boolean
