+++ plugin/src/main/kotlin/com/vanniktech/maven/publish/internal/central/MavenCentralProject.kt
@@ -1,4 +1,4 @@
-package com.vanniktech.maven.publish.central
+package com.vanniktech.maven.publish.internal.central
 
 import java.io.File
